+++ app/Services/Ftp.php
@@ -92,7 +92,6 @@ class Ftp extends Directory_Listing_Base implements Service {
 	public function init(): void {
 		$this->title = __( 'Choose file(s) from a FTP server', 'external-files-in-media-library' );
 		add_filter( 'efml_directory_listing_objects', array( $this, 'add_directory_listing' ) );
-		add_filter( 'eml_add_dialog', array( $this, 'add_option_for_local_import' ), 10, 2 );
 		add_filter( 'efml_service_ftp_hide_file', array( $this, 'prevent_not_allowed_files' ), 10, 4 );
 	}
 
@@ -106,27 +105,6 @@ class Ftp extends Directory_Listing_Base implements Service {
 	public function add_directory_listing( array $directory_listing_objects ): array {
 		$directory_listing_objects[] = $this;
 		return $directory_listing_objects;
-	}
-
-	/**
-	 * Add option to import from local directory.
-	 *
-	 * @param array<string,mixed> $dialog The dialog.
-	 * @param array<string,mixed> $settings The requested settings.
-	 *
-	 * @return array<string,mixed>
-	 */
-	public function add_option_for_local_import( array $dialog, array $settings ): array {
-		// bail if "no_services" is set in settings.
-		if ( isset( $settings['no_services'] ) ) {
-			return $dialog;
-		}
-
-		// add the hint for local import.
-		$dialog['texts'][] = '<details><summary>' . __( 'Or add from FTP-server directory', 'external-files-in-media-library' ) . '</summary><div><label for="eml_ftp"><a href="' . Directory_Listing::get_instance()->get_view_directory_url( $this ) . '" class="button button-secondary">' . esc_html__( 'Add from FTP server directory', 'external-files-in-media-library' ) . '</a></label></div></details>';
-
-		// return resulting dialog.
-		return $dialog;
 	}
 
 	/**